requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
pillow==10.2.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
qrcode==7.4.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
